+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 16.0.0
+**Implemented New Features and Enhancements:**
+- Update themes
+- Upgrade to PrimeNG 16
+- Upgrade to Angular 16
+
 # 15.1.0
 
 - General style adjustments